requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
constantly==23.10.4
contourpy==1.0.7
cryptography==41.0.7
cycler==0.11.0
daphne==4.0.0
debugpy==1.6.6
decorator==5.1.1
distro==1.8.0
Django==4.2.6
django-debug-toolbar==4.2.0
django-storages==1.14.2
entrypoints==0.4
executing==1.2.0
fastapi==0.104.1
ffmpy==0.3.0
filelock==3.10.7
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.10.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
gradio==4.7.1
gradio_client==0.7.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.19.4
hyperlink==21.0.0
idna==3.4
importlib-resources==6.1.1
incremental==22.10.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nest-asyncio==1.5.6
numpy==1.24.2
openai==0.28.0
orjson==3.8.9
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.14.0
PyMySQL==1.1.0
pyOpenSSL==23.3.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
redis==5.0.1
requests==2.28.2
rfc3986==1.5.0
rich==13.7.0
rsa==4.9
semantic-version==2.10.0
service-identity==23.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
tomlkit==0.12.0
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==23.10.0
txaio==23.1.1
typer==0.9.0
typing_extensions==4.8.0
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websockets==11.0
yarl==1.8.2
zope.interface==6.1
